Refactor of tests
