+++ server/src/test/seed/seed.js
@@ -1,6 +1,8 @@
 'use strict'
 
 const {ObjectId} = require('mongodb')
+const jwt = require('jsonwebtoken')
+const config = require('config')
 
 const User = require('./../../models/User')
 const Device = require('./../../models/Device')
@@ -111,4 +113,22 @@ const populateLogs = (done) => {
     .then(() => done())
 }
 
-module.exports = {users, devices, logs, populateUsers, populateDevices, populateLogs}
+const populateTokens = () => {
+  const tokenAdmin = jwt.sign({_id: users[2]._id.toHexString(), privileges: users[2].privileges},
+                                  config.get('Settings.JWT.secret'), {algorithm: config.get('Settings.JWT.algorithm'), expiresIn: config.get('Settings.JWT.expiration'), issuer: config.get('Settings.JWT.issuer')}
+                                  ).toString()
+
+  const notAuthorizedToken = jwt.sign({_id: users[0]._id.toHexString(), privileges: users[0].privileges},
+                                  config.get('Settings.JWT.secret'), {algorithm: config.get('Settings.JWT.algorithm'), expiresIn: config.get('Settings.JWT.expiration'), issuer: config.get('Settings.JWT.issuer')}
+                                  ).toString()
+  const authorizedToken = jwt.sign({_id: users[1]._id.toHexString(), privileges: users[1].privileges},
+                                  config.get('Settings.JWT.secret'), {algorithm: config.get('Settings.JWT.algorithm'), expiresIn: config.get('Settings.JWT.expiration'), issuer: config.get('Settings.JWT.issuer')}
+                                  ).toString()
+  const nonExistentId = new ObjectId()
+  const nonExistentUserToken = jwt.sign({_id: nonExistentId.toHexString(), privileges: users[2].privileges},
+                                                              config.get('Settings.JWT.secret'), {algorithm: config.get('Settings.JWT.algorithm'), expiresIn: config.get('Settings.JWT.expiration'), issuer: config.get('Settings.JWT.issuer')}
+                                                              ).toString()
+  return {tokenAdmin, notAuthorizedToken, authorizedToken, nonExistentUserToken}
+}
+
+module.exports = {users, devices, logs, populateUsers, populateDevices, populateLogs, populateTokens}
